PRAC-2: Finished 1-3 counter example design road network shortest time.

+++ src/main/java/org/example/roadnetwork/Distance.java
@@ -8,10 +8,10 @@ public interface Distance {
     int distanceToInMiles();
 
     /**
-     * @return The time cost(in minutes) travelling to this place
+     * @return The mph time cost travelling to this place
      *         due to external factors.
      *         e.g traffic stops, car congestion, accidents etc.
      */
-    int timeCostPerMile();
+    float timeCostPerMile();
 
 }
