+++ src/main/java/org/example/roadnetwork/model/Journey.java
@@ -5,9 +5,9 @@ import org.example.roadnetwork.Distance;
 public class Journey implements Distance {
 
     private int distanceToPointInMiles = 0;
-    private int timeCostPerMile = 0;
+    private float timeCostPerMile = 0;
 
-    public Journey(final int distanceTo, final int timeCost) {
+    public Journey(final int distanceTo, final float timeCost) {
         this.distanceToPointInMiles = distanceTo;
         this.timeCostPerMile = timeCost;
     }
@@ -18,7 +18,7 @@ public class Journey implements Distance {
     }
 
     @Override
-    public int timeCostPerMile() {
+    public float timeCostPerMile() {
         return this.timeCostPerMile;
     }
